requirements.txt
bcrypt==3.2.0
beautifulsoup4==4.10.0
bigquery==0.0.20
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.1
dacktool==0.0.7
dbstream==0.1.13
google-api-core==2.3.2
google-api-python-client==1.7.11
google-auth==2.6.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.6
google-cloud-bigquery==2.32.0
google-cloud-bigquery-storage==2.11.0
google-cloud-core==2.2.2
google-cloud-secret-manager==2.7.2
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
googleauthentication==0.0.16
grpc-google-iam-v1==0.12.3
grpcio==1.43.0
grpcio-status==1.43.0
httplib2==0.20.4
idna==3.3
libcst==0.4.1
mypy-extensions==0.4.3
numpy==1.22.2
oauthlib==3.2.0
packaging==21.3
pandas==1.4.0
paramiko==2.9.2
proto-plus==1.19.9
protobuf==3.19.4
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
soupsieve==2.3.1
sshtunnel==0.1.5
typing-inspect==0.7.1
typing_extensions==4.0.1
uritemplate==3.0.1
urllib3==1.26.8
